locale update & entity validation update (pls dont break)

+++ client/client.lua
@@ -20,8 +20,11 @@ AddEventHandler('jd-headbag:downstream', function()
         HeadbagEntity = CreateObject(GetHashKey("prop_headbag"), 0, 0, 0, true, true, true)
         AttachEntityToEntity(HeadbagEntity, PlayerPedId(), GetPedBoneIndex(PlayerPedId(), 12844), 0.26, 0.02, 0, 0, 270.0, 75.0, true, true, false, true, 1, true)
     else
-        DeleteEntity(HeadbagEntity)
-        SetEntityAsNoLongerNeeded(HeadbagEntity)
+        if HeadbagEntity and DoesEntityExist(HeadbagEntity) then
+            DeleteEntity(HeadbagEntity)
+            SetEntityAsNoLongerNeeded(HeadbagEntity)
+        end
+        HeadbagEntity = nil
     end
 end)
 
@@ -105,7 +108,7 @@ exports('GetHeadbagStatus', GetHeadbagStatus)
 
 AddEventHandler('onResourceStop', function(resourceName)
     if resourceName == GetCurrentResourceName() then
-        if HeadbagEntity ~= nil then
+        if HeadbagEntity and DoesEntityExist(HeadbagEntity) then
             DeleteEntity(HeadbagEntity)
             SetEntityAsNoLongerNeeded(HeadbagEntity)
         end
